+++ src/components/user/UserPreOrderTab.tsx
@@ -7,7 +7,7 @@ interface UserOrderTabProps {
 
 const UserPreOrderTab = ({ activeTab, setActiveTab }: UserOrderTabProps) => {
   return (
-    <div className="flex justify-center w-full">
+    <div className='flex justify-center w-full'>
       <div
         className={classNames(
           'flex justify-center w-full items-center py-2 cursor-pointer font-bold',
